Switch to criterion and benchmarks scaffolding
