+++ Cargo.toml
@@ -6,8 +6,17 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+criterion = "0.4.0"
 fs = "0.0.5"
 geo = "0.23.1"
 geo-types = { version = "0.7.8", features = ["use-rstar_0_9"] } 
 rstar = "0.9.3"
 wkt = "0.10.3"
+
+[profile.release]
+lto = true
+codegen-units = 1
+
+[[bench]]
+name = "benchmark"
+harness = false
